+++ src/guzzle/src/HandlerStackFactory.php
@@ -56,13 +56,7 @@ class HandlerStackFactory
         $middlewares = array_merge($this->middlewares, $middlewares);
 
         if (Coroutine::getCid() > 0) {
-            if ($this->usePoolHandler) {
-                $handler = make(PoolHandler::class, [
-                    'option' => $option,
-                ]);
-            } else {
-                $handler = new CoroutineHandler();
-            }
+            $handler = $this->getHandler($option);
         }
 
         $stack = HandlerStack::create($handler);
@@ -79,5 +73,20 @@ class HandlerStackFactory
         }
 
         return $stack;
+    }
+
+    protected function getHandler(array $option)
+    {
+        if ($this->usePoolHandler) {
+            return make(PoolHandler::class, [
+                'option' => $option,
+            ]);
+        }
+
+        if (class_exists(ApplicationContext::class)) {
+            return make(CoroutineHandler::class);
+        }
+
+        return new CoroutineHandler();
     }
 }